1.1.7

+++ wecom-objects/src/main/java/cn/felord/domain/corpay/miniapppay/GoodsDetailItem.java
@@ -1,3 +1,18 @@
+/*
+ *  Copyright (c) 2023. felord.cn
+ *    Licensed under the Apache License, Version 2.0 (the "License");
+ *    you may not use this file except in compliance with the License.
+ *  You may obtain a copy of the License at
+ *       https://www.apache.org/licenses/LICENSE-2.0
+ *  Website:
+ *       https://felord.cn
+ *  Unless required by applicable law or agreed to in writing, software
+ *  distributed under the License is distributed on an "AS IS" BASIS,
+ *  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ *  See the License for the specific language governing permissions and
+ *  limitations under the License.
+ */
+
 package cn.felord.domain.corpay.miniapppay;
 
 import com.fasterxml.jackson.annotation.JsonCreator;
@@ -14,7 +29,7 @@ import lombok.ToString;
 public class GoodsDetailItem {
     private final String merchantGoodsId;
     private final Integer quantity;
-    private final Integer unitPrice;
+    private final Long unitPrice;
     private final String goodsName;
     private final String wechatpayGoodsId;
 
@@ -25,7 +40,7 @@ public class GoodsDetailItem {
      * @param quantity        the quantity
      * @param unitPrice       the unit price
      */
-    public GoodsDetailItem(String merchantGoodsId, int quantity, int unitPrice) {
+    public GoodsDetailItem(String merchantGoodsId, int quantity, long unitPrice) {
         this(merchantGoodsId, quantity, unitPrice, null);
     }
 
@@ -37,7 +52,7 @@ public class GoodsDetailItem {
      * @param unitPrice       the unit price
      * @param goodsName       the goods name
      */
-    public GoodsDetailItem(String merchantGoodsId, int quantity, int unitPrice, String goodsName) {
+    public GoodsDetailItem(String merchantGoodsId, int quantity, long unitPrice, String goodsName) {
         this(merchantGoodsId, quantity, unitPrice, goodsName, null);
     }
 
@@ -53,7 +68,7 @@ public class GoodsDetailItem {
     @JsonCreator
     public GoodsDetailItem(@JsonProperty("merchant_goods_id") String merchantGoodsId,
                            @JsonProperty("quantity") int quantity,
-                           @JsonProperty("unit_price") int unitPrice,
+                           @JsonProperty("unit_price") long unitPrice,
                            @JsonProperty("goods_name") String goodsName,
                            @JsonProperty("wechatpay_goods_id") String wechatpayGoodsId) {
         this.merchantGoodsId = merchantGoodsId;
